Fix incorrect jsons
