+++ common/src/main/kotlin/entity/DiscordChannel.kt
@@ -176,6 +176,39 @@ class DiscordThreadMetadata(
     @SerialName("archive_timestamp")
     val archiveTimestamp: String,
     @SerialName("auto_archive_duration")
-    val autoArchiveDuration: Int,
+    val autoArchiveDuration: ArchiveDuration,
     val locked: OptionalBoolean = OptionalBoolean.Missing
 )
+
+@Serializable(with = ArchiveDuration.Serializer::class)
+sealed class ArchiveDuration(val duration: Int) {
+    class Unknown(duration: Int) : ArchiveDuration(duration)
+    object Hour : ArchiveDuration(60)
+    object Day : ArchiveDuration(1440)
+    object ThreeDays : ArchiveDuration(4320)
+    object Week : ArchiveDuration(10080)
+
+    object Serializer : KSerializer<ArchiveDuration> {
+        override fun deserialize(decoder: Decoder): ArchiveDuration {
+            val value = decoder.decodeInt()
+            return values.firstOrNull { it.duration == value } ?: Unknown(value)
+        }
+
+        override val descriptor: SerialDescriptor
+            get() = PrimitiveSerialDescriptor("AutoArchieveDuration", PrimitiveKind.INT)
+
+        override fun serialize(encoder: Encoder, value: ArchiveDuration) {
+            encoder.encodeInt(value.duration)
+        }
+    }
+
+    companion object {
+        val values: Set<ArchiveDuration>
+            get() = setOf(
+                Hour,
+                Day,
+                ThreeDays,
+                Week,
+            )
+    }
+}
